pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tasksync"
version = "0.1.0"
description = "Sync Google Tasks with Todoist"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["google-tasks", "todoist", "sync", "tasks"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.80.0",
    "requests>=2.28.0",
    "python-dotenv>=0.21.0",
    "click>=8.1.0",
    "pydantic>=1.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=5.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "tox>=3.28.0",
]

[project.scripts]
tasksync = "tasksync.cli:main"

[tool.setuptools]
packages = ["tasksync"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
